Refactor collision detection

+++ src/Util.cpp
@@ -207,7 +207,7 @@ GetFileDirectory(char *FilePath, i32 FilePathCount,
 void
 PrintVBODataF(u32 VBO, size_t FloatCount)
 {
-    printf("VBO %d data (%d floats):\n", VBO, FloatCount);
+    printf("VBO %d data (%d floats):\n", VBO, (i32) FloatCount);
     glBindBuffer(GL_ARRAY_BUFFER, VBO);
     f32 *DebugVBOData = (f32 *) malloc(FloatCount * sizeof(f32));
     glGetBufferSubData(GL_ARRAY_BUFFER, 0, FloatCount * sizeof(f32), DebugVBOData);
